requirements.txt
fastapi~=0.116.1
pydantic~=2.11.7
pandas~=2.3.1
numpy~=2.2.6
PyYAML~=6.0.2
pgmpy~=1.0.0
SQLAlchemy~=2.0.42
uvicorn~=0.35.0